requirements.txt
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0

streamlit>=1.32.0

groq>=0.5.0

sentence-transformers>=2.7.0
faiss-cpu>=1.8.0

requests>=2.31.0
beautifulsoup4>=4.12.0

fpdf2>=2.7.9
torchvision>=0.16.0

pandas>=2.0.0

python-dotenv>=1.0.0
 
langgraph>=0.2.0